+++ posleurokov-app/src/components/shared/List/List.module.scss
@@ -1,6 +1,6 @@
 @import 'src/assets/styles/_variables.scss';
 
-.cities {
+.list {
   margin-top: 19px;
   display: flex;
   text-decoration: none;
