+++ posleurokov-app/src/components/shared/index.js
@@ -7,4 +7,4 @@ export { Select } from './Select/Select';
 export { Sheet } from './Sheet/Sheet';
 export { Categories } from './Categories/Categories';
 export { Category } from './Category/Category';
-export { CitiesList } from './CitiesList/CitiesList';
+export { List } from './List/List';
